requirements.txt
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
imageio==2.36.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
PyMatting==1.1.13
referencing==0.35.1
rembg==2.0.61
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.15.1
tifffile==2025.1.10
tkinterdnd2==0.4.2
tqdm==4.67.1
urllib3==2.3.0
